Fix performance drop for large sizes.
Some parameters and thresholds were still using ints. Changed them to
use size_t instead.

+++ src/include/common_coll.h
@@ -79,7 +79,7 @@ static __inline__ int ncclTypeSize(ncclDataType_t type) {
 }
 
 static ncclResult_t saveKernel(int coll, const void* sendbuff, void* recvbuff, size_t count,
-    ncclDataType_t dtype, ncclRedOp_t op, int root, ncclComm_t comm, cudaStream_t stream, int nbytes) {
+    ncclDataType_t dtype, ncclRedOp_t op, int root, ncclComm_t comm, cudaStream_t stream, size_t nbytes) {
   int llMode = nbytes <= comm->llThreshold ? 1 : 0;
   int nBlocks = llMode ? 1 : LIMIT_NRINGS(nbytes, comm->nRings);
   int nThreads = llMode ? LL_NTHREADS : comm->nThreads+1;
